moved Net::Amazon::S3 to the end
Net::Amazon::S3 tends to fail the first time through installing but works fine the second time.  I suspect there is a missing dependency ( sorry, I'm not going to hunt it down [ though it's likely related to Class::MethodMaker ] )   I moved it to the end so that it may be more likely to succeed.